Partially fix double promotion.

+++ dlls/scripted.cpp
@@ -136,10 +136,10 @@ void CCineMonster::Spawn( void )
 	if( FStringNull( pev->targetname ) || !FStringNull( m_iszIdle ) )
 	{
 		SetThink( &CCineMonster::CineThink );
-		pev->nextthink = gpGlobals->time + 1.0;
+		pev->nextthink = gpGlobals->time + 1.0f;
 		// Wait to be used?
 		if( pev->targetname )
-			m_startTime = gpGlobals->time + 1E6;
+			m_startTime = gpGlobals->time + (float)1E6;
 	}
 	if( pev->spawnflags & SF_SCRIPT_NOINTERRUPT )
 		m_interruptable = FALSE;
@@ -185,7 +185,7 @@ void CCineMonster::Use( CBaseEntity *pActivator, CBaseEntity *pCaller, USE_TYPE 
 		if( pTarget->m_scriptState == SCRIPT_PLAYING )
 			return;
 
-		m_startTime = gpGlobals->time + 0.05;
+		m_startTime = gpGlobals->time + 0.05f;
 	}
 	else
 	{
@@ -356,7 +356,7 @@ void CCineMonster::PossessEntity( void )
 			pTarget->pev->effects |= EF_NOINTERP;
 			pTarget->pev->angles.y = pev->angles.y;
 			pTarget->m_scriptState = SCRIPT_WAIT;
-			m_startTime = gpGlobals->time + 1E6;
+			m_startTime = gpGlobals->time + (float)1E6;
 			// UNDONE: Add a flag to do this so people can fixup physics after teleporting monsters
 			//			pTarget->pev->flags &= ~FL_ONGROUND;
 			break;
@@ -424,7 +424,7 @@ void CCineAI::PossessEntity( void )
 			pTarget->pev->effects |= EF_NOINTERP;
 			pTarget->pev->angles.y = pev->angles.y;
 			pTarget->m_scriptState = SCRIPT_WAIT;
-			m_startTime = gpGlobals->time + 1E6;
+			m_startTime = gpGlobals->time + (float)1E6;
 			// UNDONE: Add a flag to do this so people can fixup physics after teleporting monsters
 			pTarget->pev->flags &= ~FL_ONGROUND;
 			break;
@@ -466,7 +466,7 @@ void CCineMonster::CineThink( void )
 	{
 		CancelScript();
 		ALERT( at_aiconsole, "script \"%s\" can't find monster \"%s\"\n", STRING( pev->targetname ), STRING( m_iszEntity ) );
-		pev->nextthink = gpGlobals->time + 1.0;
+		pev->nextthink = gpGlobals->time + 1.0f;
 	}
 }
 
@@ -544,7 +544,7 @@ void CCineMonster::SequenceDone( CBaseMonster *pMonster )
 	if( !( pev->spawnflags & SF_SCRIPT_REPEATABLE ) )
 	{
 		SetThink( &CBaseEntity::SUB_Remove );
-		pev->nextthink = gpGlobals->time + 0.1;
+		pev->nextthink = gpGlobals->time + 0.1f;
 	}
 
 	// This is done so that another sequence can take over the monster when triggered by the first
@@ -713,7 +713,7 @@ void CCineMonster::DelayStart( int state )
 			{
 				pTarget->m_iDelay--;
 				if( pTarget->m_iDelay <= 0 )
-					pTarget->m_startTime = gpGlobals->time + 0.05;
+					pTarget->m_startTime = gpGlobals->time + 0.05f;
 			}
 		}
 		pentCine = FIND_ENTITY_BY_TARGETNAME( pentCine, STRING( pev->targetname ) );
@@ -836,7 +836,7 @@ BOOL CBaseMonster::CineCleanup()
 			// UNDONE: ugly hack.  Don't move monster if they don't "seem" to move
 			// this really needs to be done with the AX,AY,etc. flags, but that aren't consistantly
 			// being set, so animations that really do move won't be caught.
-			if( ( oldOrigin - new_origin).Length2D() < 8.0 )
+			if( ( oldOrigin - new_origin).Length2D() < 8.0f )
 				new_origin = oldOrigin;
 
 			pev->origin.x = new_origin.x;
@@ -974,7 +974,7 @@ void CScriptedSentence::KeyValue( KeyValueData *pkvd )
 	}
 	else if( FStrEq( pkvd->szKeyName, "volume" ) )
 	{
-		m_flVolume = atof( pkvd->szValue ) * 0.1;
+		m_flVolume = atof( pkvd->szValue ) * 0.1f;
 		pkvd->fHandled = TRUE;
 	}
 	else if( FStrEq( pkvd->szKeyName, "listener" ) )
@@ -1004,7 +1004,7 @@ void CScriptedSentence::Spawn( void )
 	if( !pev->targetname )
 	{
 		SetThink( &CScriptedSentence::FindThink );
-		pev->nextthink = gpGlobals->time + 1.0;
+		pev->nextthink = gpGlobals->time + 1.0f;
 	}
 
 	switch( pev->impulse )
@@ -1030,8 +1030,8 @@ void CScriptedSentence::Spawn( void )
 	pev->impulse = 0;
 
 	// No volume, use normal
-	if( m_flVolume <= 0 )
-		m_flVolume = 1.0;
+	if( m_flVolume <= 0.0f )
+		m_flVolume = 1.0f;
 }
 
 void CScriptedSentence::FindThink( void )
@@ -1050,7 +1050,7 @@ void CScriptedSentence::FindThink( void )
 	else
 	{
 		//ALERT( at_console, "%s: can't find monster %s\n", STRING( m_iszSentence ), STRING( m_iszEntity ) );
-		pev->nextthink = gpGlobals->time + m_flRepeat + 0.5;
+		pev->nextthink = gpGlobals->time + m_flRepeat + 0.5f;
 	}
 }
 
@@ -1058,7 +1058,7 @@ void CScriptedSentence::DelayThink( void )
 {
 	m_active = TRUE;
 	if( !pev->targetname )
-		pev->nextthink = gpGlobals->time + 0.1;
+		pev->nextthink = gpGlobals->time + 0.1f;
 	SetThink( &CScriptedSentence::FindThink );
 }
 
@@ -1146,7 +1146,7 @@ BOOL CScriptedSentence::StartSentence( CBaseMonster *pTarget )
 	}
 
 	pTarget->PlayScriptedSentence( STRING( m_iszSentence ), m_flDuration,  m_flVolume, m_flAttenuation, bConcurrent, pListener );
-	ALERT( at_aiconsole, "Playing sentence %s (%.1f)\n", STRING( m_iszSentence ), m_flDuration );
+	ALERT( at_aiconsole, "Playing sentence %s (%.1f)\n", STRING( m_iszSentence ), (double)m_flDuration );
 	SUB_UseTargets( NULL, USE_TOGGLE, 0 );
 	return TRUE;
 }
@@ -1192,7 +1192,7 @@ void CFurniture::Spawn()
 	pev->sequence = 0;
 	pev->frame = 0;
 
-	//pev->nextthink += 1.0;
+	//pev->nextthink += 1.0f;
 	//SetThink( &WalkMonsterDelay );
 
 	ResetSequenceInfo();
